+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
-
+### Fixed
+- Correctly detects location of `.env` file in the current working directory.
 
 
 ## [v0.1.0] - 2025-05-26