+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - Correctly detects location of `.env` file in the current working directory.
+- Disables logging from other packages to avoid cluttering the output.
 
 
 ## [v0.1.0] - 2025-05-26