+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Added
+- Now automatically stores documents in version-indexed files, when the file already
+  exists.
+
 ### Fixed
 - Correctly detects location of `.env` file in the current working directory.
 - Disables logging from other packages to avoid cluttering the output.